feat(authentication pkg): Update Claims struct to include UserId (to store UserId as a jwt claim) and create DecodeJWT method to retrieve claims
